+++ Kumi.Game/Screens/Edit/EditorScreenStack.cs
@@ -1,0 +1,7 @@
+﻿using osu.Framework.Screens;
+
+namespace Kumi.Game.Screens.Edit;
+
+public partial class EditorScreenStack : ScreenStack
+{
+}
